add specific routers for `popup` and `connector` contexts
