+++ src/extension/connector/main.ts
@@ -1,7 +1,8 @@
 import { createApp } from "vue";
 import App from "./App.vue";
+import { router } from "./router";
 import { setupApp } from "@/boot/commonSetup";
 import { registerRpcHooks } from "@/rpc/uiRpcHandlers";
 
 registerRpcHooks();
-setupApp(createApp(App)).mount("#app");
+setupApp(createApp(App)).use(router).mount("#app");
